refactor: group and rename entities
