+++ libs/tokens/src/hooks/lists/useAddList.ts
@@ -1,0 +1,6 @@
+import { useSetAtom } from 'jotai'
+import { addListAtom } from '../../state/tokenLists/tokenListsActionsAtom'
+
+export function useAddList() {
+  return useSetAtom(addListAtom)
+}
