+++ libs/tokens/src/hooks/lists/useAllListsList.ts
@@ -1,6 +1,6 @@
 import { useAtomValue } from 'jotai'
-import { listsStatesListAtom } from '../state/tokenLists/tokenListsStateAtom'
-import { ListState } from '../types'
+import { listsStatesListAtom } from '../../state/tokenLists/tokenListsStateAtom'
+import { ListState } from '../../types'
 
 export function useAllListsList(): ListState[] {
   return useAtomValue(listsStatesListAtom)
