+++ libs/tokens/src/hooks/lists/useListsEnabledState.ts
@@ -1,6 +1,6 @@
 import { useAtomValue } from 'jotai'
-import { listsEnabledStateAtom } from '../state/tokenLists/tokenListsStateAtom'
-import { ListsEnabledState } from '../types'
+import { listsEnabledStateAtom } from '../../state/tokenLists/tokenListsStateAtom'
+import { ListsEnabledState } from '../../types'
 
 export function useListsEnabledState(): ListsEnabledState {
   return useAtomValue(listsEnabledStateAtom)
