+++ libs/tokens/src/hooks/lists/useRemoveList.ts
@@ -1,11 +1,11 @@
 import { useSetAtom } from 'jotai'
 import { removeListAnalytics } from '@cowprotocol/analytics'
 
-import { removeListAtom } from '../state/tokenLists/tokenListsActionsAtom'
-import { ListState } from '../types'
-import { getTokenListSource } from '../utils/getTokenListSource'
+import { removeListAtom } from '../../state/tokenLists/tokenListsActionsAtom'
+import { ListState } from '../../types'
+import { getTokenListSource } from '../../utils/getTokenListSource'
 
-export function useRemoveTokenList() {
+export function useRemoveList() {
   const removeList = useSetAtom(removeListAtom)
 
   return (list: ListState) => {
