+++ libs/tokens/src/hooks/tokens/favourite/useFavouriteTokens.ts
@@ -1,5 +1,5 @@
 import { useAtomValue } from 'jotai'
-import { favouriteTokensListAtom } from '../state/tokens/favouriteTokensAtom'
+import { favouriteTokensListAtom } from '../../../state/tokens/favouriteTokensAtom'
 import { TokenWithLogo } from '@cowprotocol/common-const'
 
 export function useFavouriteTokens(): TokenWithLogo[] {
