+++ libs/tokens/src/hooks/tokens/favourite/useResetFavouriteTokens.ts
@@ -1,5 +1,5 @@
 import { useSetAtom } from 'jotai'
-import { resetFavouriteTokensAtom } from '../state/tokens/favouriteTokensAtom'
+import { resetFavouriteTokensAtom } from '../../../state/tokens/favouriteTokensAtom'
 
 export function useResetFavouriteTokens() {
   return useSetAtom(resetFavouriteTokensAtom)
