+++ libs/tokens/src/hooks/tokens/favourite/useToggleFavouriteToken.ts
@@ -1,5 +1,5 @@
 import { useSetAtom } from 'jotai'
-import { toggleFavouriteTokenAtom } from '../state/tokens/favouriteTokensAtom'
+import { toggleFavouriteTokenAtom } from '../../../state/tokens/favouriteTokensAtom'
 
 export function useToggleFavouriteToken() {
   return useSetAtom(toggleFavouriteTokenAtom)
