+++ libs/tokens/src/hooks/tokens/unsupported/useAddUnsupportedToken.ts
@@ -1,5 +1,5 @@
 import { useSetAtom } from 'jotai'
-import { addUnsupportedTokenAtom } from '../state/tokens/unsupportedTokensAtom'
+import { addUnsupportedTokenAtom } from '../../../state/tokens/unsupportedTokensAtom'
 
 export function useAddUnsupportedToken() {
   return useSetAtom(addUnsupportedTokenAtom)
