+++ libs/tokens/src/hooks/tokens/unsupported/useRemoveUnsupportedToken.ts
@@ -1,5 +1,5 @@
 import { useSetAtom } from 'jotai'
-import { removeUnsupportedTokenAtom } from '../state/tokens/unsupportedTokensAtom'
+import { removeUnsupportedTokenAtom } from '../../../state/tokens/unsupportedTokensAtom'
 
 export function useRemoveUnsupportedToken() {
   return useSetAtom(removeUnsupportedTokenAtom)
