+++ libs/tokens/src/hooks/tokens/unsupported/useUnsupportedTokens.ts
@@ -1,5 +1,5 @@
 import { useAtomValue } from 'jotai'
-import { currentUnsupportedTokensAtom } from '../state/tokens/unsupportedTokensAtom'
+import { currentUnsupportedTokensAtom } from '../../../state/tokens/unsupportedTokensAtom'
 
 export function useUnsupportedTokens() {
   return useAtomValue(currentUnsupportedTokensAtom)
