+++ libs/tokens/src/hooks/tokens/useSearchToken.ts
@@ -1,18 +1,18 @@
 import { useMemo } from 'react'
 import { useAtomValue } from 'jotai'
-import { activeTokensAtom, inactiveTokensAtom } from '../state/tokens/tokensAtom'
+import { activeTokensAtom, inactiveTokensAtom } from '../../state/tokens/allTokensAtom'
 import { useDebounce } from '@cowprotocol/common-hooks'
 import { useWeb3React } from '@web3-react/core'
 import { isAddress } from '@cowprotocol/common-utils'
 import ms from 'ms.macro'
-import { getTokenSearchFilter } from '../utils/getTokenSearchFilter'
+import { getTokenSearchFilter } from '../../utils/getTokenSearchFilter'
 import useSWR from 'swr'
-import { searchTokensInApi } from '../services/searchTokensInApi'
+import { searchTokensInApi } from '../../services/searchTokensInApi'
 import { TokenWithLogo } from '@cowprotocol/common-const'
-import { environmentAtom } from '../state/environmentAtom'
-import { parseTokensFromApi } from '../utils/parseTokensFromApi'
-import { fetchTokenFromBlockchain } from '../utils/fetchTokenFromBlockchain'
-import { tokenWithLogoFromToken } from '../utils/tokenWithLogoFromToken'
+import { environmentAtom } from '../../state/environmentAtom'
+import { parseTokensFromApi } from '../../utils/parseTokensFromApi'
+import { fetchTokenFromBlockchain } from '../../utils/fetchTokenFromBlockchain'
+import { tokenWithLogoFromToken } from '../../utils/tokenWithLogoFromToken'
 
 const IN_LISTS_DEBOUNCE_TIME = ms`100ms`
 const IN_EXTERNALS_DEBOUNCE_TIME = ms`1s`
