+++ libs/tokens/src/hooks/tokens/useTokensByAddressMap.ts
@@ -1,4 +1,4 @@
-import { TokensByAddress, tokensByAddressAtom } from '../state/tokens/tokensAtom'
+import { TokensByAddress, tokensByAddressAtom } from '../../state/tokens/allTokensAtom'
 import { useAtomValue } from 'jotai'
 
 export function useTokensByAddressMap(): TokensByAddress {
