+++ libs/tokens/src/hooks/tokens/userAdded/useRemoveUserToken.ts
@@ -1,8 +1,8 @@
 import { TokenWithLogo } from '@cowprotocol/common-const'
 import { useSetAtom } from 'jotai'
-import { removeUserTokenAtom } from '../state/tokens/userAddedTokensAtom'
+import { removeUserTokenAtom } from '../../../state/tokens/userAddedTokensAtom'
 
-export function useRemoveTokenCallback(): (token: TokenWithLogo) => void {
+export function useRemoveUserToken(): (token: TokenWithLogo) => void {
   const removeUserToken = useSetAtom(removeUserTokenAtom)
 
   return (token: TokenWithLogo) => {
