+++ libs/tokens/src/hooks/tokens/userAdded/useResetUserTokens.ts
@@ -1,0 +1,6 @@
+import { useSetAtom } from 'jotai'
+import { resetUserTokensAtom } from '../../../state/tokens/userAddedTokensAtom'
+
+export function useResetUserTokens(): () => void {
+  return useSetAtom(resetUserTokensAtom)
+}
