+++ libs/tokens/src/hooks/tokens/userAdded/useUserAddedTokens.ts
@@ -1,5 +1,5 @@
 import { useAtomValue } from 'jotai'
-import { userAddedTokensListAtom } from '../state/tokens/userAddedTokensAtom'
+import { userAddedTokensListAtom } from '../../../state/tokens/userAddedTokensAtom'
 import { TokenWithLogo } from '@cowprotocol/common-const'
 
 export function useUserAddedTokens(): TokenWithLogo[] {
