--- libs/tokens/src/hooks/useAddCustomTokenLists.ts
@@ -1,6 +1,0 @@
-import { useSetAtom } from 'jotai'
-import { addListAtom } from '../state/tokenLists/tokenListsActionsAtom'
-
-export function useAddCustomTokenLists() {
-  return useSetAtom(addListAtom)
-}
