--- libs/tokens/src/hooks/useResetUserTokensCallback.ts
@@ -1,6 +1,0 @@
-import { useSetAtom } from 'jotai'
-import { resetUserTokenAtom } from '../state/tokens/userAddedTokensAtom'
-
-export function useResetUserTokensCallback(): () => void {
-  return useSetAtom(resetUserTokenAtom)
-}
